+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.2.2 under development
 
+- New #95: Add class for tag `Title` and method `Html::title()` (vjik)
 - New #96: Add classes for heading tags `H1-6` and methods `Html::h1()`, `Html::h2()`, `Html::h3()`, `Html::h4()`,
   `Html::h5()`, `Html::h6()` (vjik)